requirements.txt
# Use this version 
google-cloud-storage==1.42.2
